Rename file in common way
