+++ src/ggrc-client/js/plugins/persistent-notifier.js
